Move code from init to client/report/utils
